+++ frontend/src/components/common/Header.tsx
@@ -5,6 +5,7 @@ import type { MenuProps } from "antd"
 import { AppName } from "../../settings/settings"
 import { logout } from "../../app/authSlice"
 import { useDispatch } from "react-redux"
+import ThemeToggle from "../ThemeToggle"
 export default function Header() {
   const dispatch = useDispatch()
   const items: MenuProps["items"] = [
@@ -16,6 +17,10 @@ export default function Header() {
     {
       label: <Button onClick={() => dispatch(logout())}>Logout</Button>,
       key: "logout",
+    },
+    {
+      label: <ThemeToggle />,
+      key: "theme",
     },
   ]
   return <Menu items={items} selectable={false} mode="horizontal"></Menu>