+++ frontend/src/pages/HomePage.tsx
@@ -1,12 +1,12 @@
 import { Layout } from "antd"
 import React from "react"
-import ItemPage from "../components/ItemPage"
+import ItemPage from "../components/ItemTable"
 import Header from "../components/common/Header"
 import Footer from "../components/common/Footer"
 
 export default function HomePage() {
   return (
-    <Layout>
+    <Layout style={{ minHeight: "100vh" }}>
       <Header />
       <ItemPage />
       <Footer />